+++ src/js/configurationwidget.mjs
@@ -1,1 +1,5 @@
-export { ConfigurationWidget } from './overlays/configuration/configuration.mjs';
+export {
+    ANNOTATION_TOOLBAR_PERSIST_ID_FILE,
+    ANNOTATION_TOOLBAR_PERSIST_ID_PENCIL,
+    ConfigurationWidget,
+} from './overlays/configuration/configuration.mjs';